Fix bug causing factor of 2 discrepancy in interior versus limits
The interior `x ∈ (-1,1)` was missing a factor of two in the
normalization coefficient `γ` which caused the special limiting case
of (anti-)podal nodes to be abruptly a factor of two larger than the
rest of the domain.

Fixes the normalization error and adds a [loose] test to make sure
something so egregious doesn't show up again.