Prepare cov impl core for multicolumn calculations
Hoisting the load of r₂ outside of the loop seems to increase the
small numerical non-symmetry to above a 1/100 ulp (of covariance
maximum) level. This is still basically insignificant, so just increase
the default a bit for now — later detailed work on vectorization and
accuracy tweaks can try to get this lower again (if possible).