Fix handling of non 3-vector inputs to spherical functions
This was a missing requirement/check on the spherical functions, and
that translated to also implicitly allowing silently-incorrect
covariance calculations to proceed. Now an error is thrown if the
pixel pointing vector isn't a proper 3-vector.

The errors thrown are pretty deep in the call chain so it might be
good to also do a higher-level check, but for now, this at least
prevents wrong computations from ocurring.