Add `fields` metadata to reading/writing observing matrices.
Because the number of metadata fields continues to grow, I think it
makes more sense to rework the interface quite a bit and to return
just a 2-tuple of results upon reading — the observing matrix `R`
itself, and then a named tuple of all the recognized metadata which
was loaded.

The field names of the metadata structure have also been changed, with
`pixr` -> `pixels_right` and `pixl` -> `pixels_left` to make the names
more self-descriptive.