Simplify uses of at-inbounds
The use of at-propagate_inbounds is motivated by it being far easier
to write at-inbounds at the callsite than to wrap the entire function
in an inbounds block, but we also observe some (as yet not understood)
extra allocations if we take the latter approach rather than the
former.

Since we want to ensure there aren't extra allocations in when all of
the working space buffers have been preallocated, we instead just
simplify the code a bit by not calling the implementation functions
twice across both branches of a conditional. Presumably LLVM is
deduplicating the nearly identical calls (since Julia will inline
to support propagating the inbounds annotation), but this should make
the job of the compiler ever so slightly easier.