Add a function barrier to force common code path
We have been observing CI failures that indicate a numerical
difference between covariance calculations called with just a Legendre
normalization versus with a pre-allocated working space.

This is not the expected result since the computational kernel always
leads to using the working space structure; we find that we can force
agreement by adding this function barrier into the code path.