Fix allocation in pixelcovariance! with pre-allocated work space
The type constraint on the `FweightsWork` field within the
`PixelCovWork` struct was missing a type parameter, and therefore it
was constrained to be an abstract type.

Rather than adding another type parameter for the 4th param explicitly,
instead just switch to allowing the type system to choose the explicit
type of the field and just constrain it to be some kind of
`FweightsWork`.

Fixes the one broken test which was recognizing that the design should
have meant that the pixel covariance kernel is allocation-free if given
pre-allocated work space, but it wasn't.