Import StaticArrays.SVector into docs to have type-only printing
This avoids a doctest failures from mismatched owner modules depending
on the version of StaticArrays which is loaded. For StaticArrays < v1.5,
the fully-qualified type is `StaticArrays.SVector`, but starting with
v1.5 the type definition moved to `StaticArraysCore.SVector`.

By importing explicitly from StaticArrays into the environment, the
printing system skips the module component and therefore we don't see
any discrepancy in printing no matter which version of
Julia + StaticArrays is being used.